refactor(frontend): extrait les composants partagés et élimine les duplications
- Centralise typeOptions/statusOptions dans enums.ts
- Extrait getCoverSrc dans utils/coverUtils.ts
- Extrait les labels de sync dans utils/syncLabels.ts
- Crée SelectListbox réutilisable (Filters, ComicForm, MergeSeries)
- Découpe ComicForm.tsx (1180→398 lignes) en useComicForm, TomeTable,
  LookupSection, AuthorAutocomplete
- Ajoute tests unitaires pour les nouvelles utilités

fixes #169
