+++ frontend/src/pages/Trash.tsx
@@ -7,6 +7,7 @@ import SkeletonBox from "../components/SkeletonBox";
 import { usePermanentDelete, useRestoreComic, useTrash } from "../hooks/useTrash";
 import type { ComicSeries } from "../types/api";
 import { ComicTypePlaceholder } from "../types/enums";
+import { getCoverSrc } from "../utils/coverUtils";
 
 export default function Trash() {
   const { data, isLoading } = useTrash();
@@ -47,7 +48,7 @@ export default function Trash() {
               <img
                 alt={comic.title}
                 className="h-12 w-9 rounded object-cover"
-                src={comic.coverImage ? `/uploads/covers/${comic.coverImage}` : (comic.coverUrl ?? ComicTypePlaceholder[comic.type])}
+                src={getCoverSrc(comic) ?? ComicTypePlaceholder[comic.type]}
               />
               <span className="flex-1 font-medium text-text-primary">{comic.title}</span>
               <button
